:sparkles: feature(post_manager): post_manager重构完成

+++ post_manager/src/main/java/edu/hunau/entity/ForumArticle.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class ForumArticle implements Serializable {
-    private Integer articleId;
+    private Long articleId;
 
-    private Integer userId;
+    private Long userId;
 
     private String title;
 
@@ -32,19 +32,19 @@ public class ForumArticle implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getArticleId() {
+    public Long getArticleId() {
         return articleId;
     }
 
-    public void setArticleId(Integer articleId) {
+    public void setArticleId(Long articleId) {
         this.articleId = articleId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
